cleaned up

+++ Program.cs
@@ -13,10 +13,13 @@ namespace SteamVRnoASW
             sw.WriteLine("server:asw.Off");
             sw.WriteLine("exit");
             sw.Close();
+            File.SetAttributes(
+                "asw_off.txt",
+                FileAttributes.Hidden);
         }
         public static void GetCLIPath()
         {
-            // second try at getting path but more complex and actually working
+            // get oculus install path and launch cli with instructions file
             string oculusPath = String.Empty;
             RegistryKey localKey = RegistryKey.OpenBaseKey(Microsoft.Win32.RegistryHive.LocalMachine, RegistryView.Registry64);
             localKey = localKey.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\Oculus");
@@ -25,15 +28,21 @@ namespace SteamVRnoASW
                 oculusPath = localKey.GetValue("InstallLocation").ToString();
             }
             string cliPath = oculusPath + @"Support\oculus-diagnostics\OculusDebugToolCLI.exe";
-            System.Diagnostics.Process.Start(cliPath, "/C -f asw_off.txt");
+            while (File.Exists("asw_off.txt") == false)
+            {
+                System.Threading.Thread.Sleep(500);
+            }
+            if (File.Exists("asw_off.txt") == true)
+            {
+                System.Diagnostics.Process.Start(cliPath, "/C -f asw_off.txt");
+                System.Threading.Thread.Sleep(1000);
+                File.Delete(".\\asw_off.txt");
+            }  
         }
         public static void Main(string[] args)
         {
             ASWCommandFile();
-            System.Threading.Thread.Sleep(1000);
             GetCLIPath();
-            System.Threading.Thread.Sleep(1000);
-            File.Delete(".\\asw_off.txt");
             System.Diagnostics.Process.Start("cmd.exe", "/C start steam://launch/250820");
         }
     }
